The main menu and open deck menus are working just fine.
Added the config crate to handle persistent settings and using the 'most recent deck' feature.

Up next is the deck view itself, which will be a little tricky in terms of layout. Ideally it would be appropriate for the size of the terminal, but that's a significant amount of work from here. Just getting it to work as a full-screen app should be the primary goal right now.

This reorganization of the code is working wonders, by the way.