Just spent two days basically going in a circle, trying to make the DbContext work. It doesn't, and probably never will. Premature optimization kills momentum. Need to get that tattooed on the inside of my eyelids.
On the bright side, the card database has been migrated to the more updated version. Another couple of things for the tasklist; on future migrations, make the correction of meld cards automatic; add loyalty as a field of cards.

The next immediate task is to transfer all references from Cards to NewCards, then rename NewCards, then implement the screen to add a deck.