Should have committed earlier. Two major changes; the deck stat screen is in (but very early beta) and we can import decks again.
Deck import only works for bare txt files and doesn't report problems as it should; will have to work on that.

The deck stat screen is... bad. It's ugly, of course, but the data format required (a ref to vector of string slices) is so bad I can't believe it's real. Why would anyone design it this way? Anyway, I need to run a second pass on it so that it doesn't hang the whole program for minutes at a time loading all the purchase data when we want to see the stats, and that means finally dealing with async. Joy.